refactor: gather the optimizers in the subpackage optimizers
optimization.py becomes optimizers/sgd.py, and cmaes.py becomes
optimizers/cmaes.py. The search box and the objective functions of warmstart
(score, evaluate, compile_score, compile_evaluate) move to optimizers/search.py,
because CMA-ES uses them too. warm_start stays in warmstart.py for now.

Breaking: el.cmaes, el.optimization and el.warmstart.score, compile_score and
PENALTY are now under el.optimizers.
