+++ src/elicito/optimizers/__init__.py
@@ -1,0 +1,7 @@
+"""
+Optimizers that fit the prior hyperparameters
+"""
+
+from elicito.optimizers import cmaes, search, sgd
+
+__all__ = ["cmaes", "search", "sgd"]
